requirements.txt
transformers==4.37.2
datasets==2.18.0
accelerate==0.27.2
deepspeed==0.12.6
pydantic==1.10.8
numpy==1.23.5
einops
huggingface_hub
matplotlib
omegaconf
openai==1.3.5
opencv-python>=4.4.0.46
openpyxl
pandas>=1.5.3
pillow
portalocker
pycocoevalcap
python-dotenv
requests
rich
seaborn
sentencepiece
sty
tabulate
tiktoken
timeout-decorator
tqdm
typing_extensions==4.7.1
visual_genome
xlsxwriter
xtuner